+++ src/main/java/com/helloworld/data/CommentReadAction.java
@@ -28,10 +28,10 @@ public class CommentReadAction {
     private LocalDateTime timeStamp;
 
     @Column(nullable = false)
-    private Float secondsSpent;
+    private Double secondsSpent;
 
     @Column(nullable = false)
-    private Float nrOfCoins;
+    private Double nrOfCoins;
 
     public CommentReadAction() {
 
@@ -42,15 +42,15 @@ public class CommentReadAction {
         this.timeStamp = timeStamp;
     }
 
-    public void setSecondsSpent(Float secondsSpent) {
+    public void setSecondsSpent(Double secondsSpent) {
         this.secondsSpent = secondsSpent;
     }
 
-    public void setNrOfCoins(Float nrOfCoins) {
+    public void setNrOfCoins(Double nrOfCoins) {
         this.nrOfCoins = nrOfCoins;
     }
 
-    public Float getNrOfCoins() {
+    public Double getNrOfCoins() {
         return nrOfCoins;
     }
 }
