+++ src/main/java/com/helloworld/data/dto/AuthorStatsDTO.java
@@ -4,12 +4,12 @@ public class AuthorStatsDTO implements Comparable {
     public String username;
     public String fullName;
     public int nrOfArticles;
-    public float totalCoins;
+    public double totalCoins;
     public int rank;
     public String mostReadArticleName;
     public String mostReadArticleURL;
 
-    public AuthorStatsDTO(String username, String fullName, int nrOfArticles, float totalCoins, String mostReadArticleName, String mostReadArticleURL) {
+    public AuthorStatsDTO(String username, String fullName, int nrOfArticles, double totalCoins, String mostReadArticleName, String mostReadArticleURL) {
         this.username = username;
         this.fullName = fullName;
         this.nrOfArticles = nrOfArticles;
@@ -22,7 +22,7 @@ public class AuthorStatsDTO implements Comparable {
     @Override
     public int compareTo(Object o) {
         if (o instanceof AuthorStatsDTO) {
-            return -Float.compare(this.totalCoins, ((AuthorStatsDTO) o).totalCoins);
+            return -Double.compare(this.totalCoins, ((AuthorStatsDTO) o).totalCoins);
         } else {
             return 0;
         }
